fix writer
