--- core/constants.py
@@ -1,1 +1,0 @@
-INTERMEDIATE_REPLACEMENT = "--_--"
