--- core/paths.py
@@ -1,5 +1,0 @@
-from pathlib import Path
-
-
-def get_project_path() -> Path:
-    return Path(__file__).parent.parent
